Split tag_resolver into per-function modules
Mirror the image_resolver layout: one function per file, re-exported
from the package __init__. Split the test file the same way.
