--- lightly_studio/tests/resolvers/test_tag_resolver.py
@@ -1,544 +1,0 @@
-from uuid import uuid4
-
-import pytest
-from sqlalchemy.exc import IntegrityError
-from sqlmodel import Session, col, select
-
-from lightly_studio.models.sample import SampleTable
-from lightly_studio.models.tag import TagCreate
-from lightly_studio.resolvers import tag_resolver
-from tests.helpers_resolvers import create_collection, create_image, create_tag
-
-
-def test_create_tag(db_session: Session) -> None:
-    collection = create_collection(session=db_session)
-    collection_id = collection.collection_id
-    tag = create_tag(session=db_session, collection_id=collection_id, tag_name="example_tag")
-    assert tag.name == "example_tag"
-
-
-def test_remove_sample_ids_from_tag_id__exceeds_postgres_param_limit(db_session: Session) -> None:
-    # More sample ids than PostgreSQL's 65,535-parameter cap.
-    collection_id = create_collection(session=db_session).collection_id
-    tag = create_tag(session=db_session, collection_id=collection_id, tag_name="t")
-    sample_ids = [uuid4() for _ in range(70_000)]
-
-    result = tag_resolver.remove_sample_ids_from_tag_id(
-        session=db_session, tag_id=tag.tag_id, sample_ids=sample_ids
-    )
-
-    assert result is not None
-
-
-def test_create_tag__unique_tag_name(db_session: Session) -> None:
-    collection = create_collection(session=db_session)
-    collection_id = collection.collection_id
-
-    create_tag(session=db_session, collection_id=collection_id, tag_name="example_tag")
-
-    # trying to create a tag with the same name results in an IntegrityError
-    with pytest.raises(IntegrityError):
-        tag_resolver.create(
-            session=db_session,
-            tag=TagCreate(
-                collection_id=collection_id,
-                name="example_tag",
-                kind="sample",
-            ),
-        )
-    db_session.rollback()
-
-
-def test_read_tags(db_session: Session) -> None:
-    collection = create_collection(session=db_session)
-    collection_id = collection.collection_id
-
-    tag_1 = create_tag(session=db_session, collection_id=collection_id, tag_name="tag_1")
-    create_tag(session=db_session, collection_id=collection_id, tag_name="tag_2")
-    create_tag(session=db_session, collection_id=collection_id, tag_name="tag_3")
-
-    # get all tags of a collection
-    tags = tag_resolver.get_all_by_collection_id(session=db_session, collection_id=collection_id)
-    assert len(tags) == 3
-    # check order
-    tag = tags[0]
-    assert tag.tag_id == tag_1.tag_id
-    assert tag.name == tag_1.name
-
-
-def test_read_tags__paginated(db_session: Session) -> None:
-    collection = create_collection(session=db_session)
-    collection_id = collection.collection_id
-
-    total = 10
-    chunk_size = total // 2
-    for i in range(total):
-        create_tag(session=db_session, collection_id=collection_id, tag_name=f"example_tag_{i}")
-
-    # get first chunk/page
-    page_1 = tag_resolver.get_all_by_collection_id(
-        session=db_session,
-        collection_id=collection.collection_id,
-        offset=0,
-        limit=chunk_size,
-    )
-    assert len(page_1) == chunk_size
-
-    # get second chunk/page
-    page_2 = tag_resolver.get_all_by_collection_id(
-        session=db_session,
-        collection_id=collection.collection_id,
-        offset=5,
-        limit=chunk_size,
-    )
-    assert len(page_2) == chunk_size
-
-    # assert that the two chunks are different
-    assert page_1 != page_2
-    assert page_1[0].name != page_2[0].name
-
-
-def test_read_tag(db_session: Session) -> None:
-    collection = create_collection(session=db_session)
-    collection_id = collection.collection_id
-
-    tag = create_tag(session=db_session, collection_id=collection_id)
-
-    tag_read = tag_resolver.get_by_id(session=db_session, tag_id=tag.tag_id)
-    assert tag_read is not None
-    assert tag_read.tag_id == tag.tag_id
-    assert tag_read.name == "example_tag"
-
-
-def test_rename_tag(db_session: Session) -> None:
-    collection = create_collection(session=db_session)
-    collection_id = collection.collection_id
-
-    tag = create_tag(session=db_session, collection_id=collection_id)
-
-    tag_renamed = tag_resolver.rename(session=db_session, tag_id=tag.tag_id, new_name="updated_tag")
-    assert tag_renamed is not None
-    assert tag_renamed.name == "updated_tag"
-
-
-def test_rename_tag__preserves_sample_links(db_session: Session) -> None:
-    collection = create_collection(session=db_session)
-    collection_id = collection.collection_id
-
-    tag = create_tag(session=db_session, collection_id=collection_id)
-    image_1 = create_image(
-        session=db_session, collection_id=collection_id, file_path_abs="sample1.png"
-    )
-    image_2 = create_image(
-        session=db_session, collection_id=collection_id, file_path_abs="sample2.png"
-    )
-
-    tag_resolver.add_tag_to_sample(session=db_session, tag_id=tag.tag_id, sample=image_1.sample)
-    tag_resolver.add_tag_to_sample(session=db_session, tag_id=tag.tag_id, sample=image_2.sample)
-
-    tag_renamed = tag_resolver.rename(session=db_session, tag_id=tag.tag_id, new_name="updated_tag")
-
-    assert tag_renamed is not None
-    assert tag_renamed.name == "updated_tag"
-    assert tag_renamed.tag_id == tag.tag_id
-    assert sorted(sample.sample_id for sample in tag_renamed.samples) == sorted(
-        [image_1.sample.sample_id, image_2.sample.sample_id]
-    )
-
-
-def test_rename_tag__unique_tag_name(db_session: Session) -> None:
-    collection = create_collection(session=db_session)
-    collection_id = collection.collection_id
-
-    tag_1 = create_tag(session=db_session, collection_id=collection_id, tag_name="example_tag_1")
-    tag_2 = create_tag(session=db_session, collection_id=collection_id, tag_name="some_other_tag")
-
-    # renaming a tag to an existing name results in an IntegrityError
-    with pytest.raises(IntegrityError):
-        tag_resolver.rename(
-            session=db_session,
-            tag_id=tag_1.tag_id,
-            new_name=tag_2.name,
-        )
-    db_session.rollback()
-
-
-def test_rename_tag__unique_tag_name__preserves_original_tag_and_links(
-    db_session: Session,
-) -> None:
-    collection = create_collection(session=db_session)
-    collection_id = collection.collection_id
-
-    tag_1 = create_tag(session=db_session, collection_id=collection_id, tag_name="example_tag_1")
-    tag_2 = create_tag(session=db_session, collection_id=collection_id, tag_name="some_other_tag")
-    image_1 = create_image(
-        session=db_session, collection_id=collection_id, file_path_abs="sample1.png"
-    )
-    image_2 = create_image(
-        session=db_session, collection_id=collection_id, file_path_abs="sample2.png"
-    )
-
-    tag_resolver.add_tag_to_sample(session=db_session, tag_id=tag_1.tag_id, sample=image_1.sample)
-    tag_resolver.add_tag_to_sample(session=db_session, tag_id=tag_1.tag_id, sample=image_2.sample)
-
-    with pytest.raises(IntegrityError):
-        tag_resolver.rename(
-            session=db_session,
-            tag_id=tag_1.tag_id,
-            new_name=tag_2.name,
-        )
-    db_session.rollback()
-
-    original_tag = tag_resolver.get_by_id(session=db_session, tag_id=tag_1.tag_id)
-    assert original_tag is not None
-    assert original_tag.name == "example_tag_1"
-    assert sorted(sample.sample_id for sample in original_tag.samples) == sorted(
-        [image_1.sample.sample_id, image_2.sample.sample_id]
-    )
-
-
-def test_rename_tag__unique_tag_name__different_kind(
-    db_session: Session,
-) -> None:
-    collection = create_collection(session=db_session)
-    collection_id = collection.collection_id
-
-    sample_tag = create_tag(
-        session=db_session,
-        collection_id=collection_id,
-        kind="sample",
-        tag_name="sample_tag_1",
-    )
-    annotation_tag = create_tag(
-        session=db_session,
-        collection_id=collection_id,
-        kind="annotation",
-        tag_name="annotation_tag_1",
-    )
-
-    # renaming a tag to an existing name but for a different kind is allowed
-    tag_renamed = tag_resolver.rename(
-        session=db_session,
-        tag_id=sample_tag.tag_id,
-        new_name=annotation_tag.name,
-    )
-    assert tag_renamed is not None
-    assert tag_renamed.name == annotation_tag.name
-
-
-def test_rename_tag__unknown_tag_returns_none(db_session: Session) -> None:
-    collection = create_collection(session=db_session)
-    collection_id = collection.collection_id
-
-    create_tag(session=db_session, collection_id=collection_id)
-
-    tag_renamed = tag_resolver.rename(
-        session=db_session,
-        tag_id=uuid4(),
-        new_name="unknown_tag",
-    )
-    assert tag_renamed is None
-
-
-def test_delete_tag(db_session: Session) -> None:
-    collection = create_collection(session=db_session)
-    collection_id = collection.collection_id
-
-    tag = create_tag(session=db_session, collection_id=collection_id)
-
-    # Delete the tag
-    tag_resolver.delete(session=db_session, tag_id=tag.tag_id)
-
-    # assert tag was deleted
-    tag_deleted = tag_resolver.get_by_id(session=db_session, tag_id=tag.tag_id)
-    assert tag_deleted is None
-
-
-def test_delete_tag__removes_sample_links(db_session: Session) -> None:
-    collection = create_collection(session=db_session)
-    collection_id = collection.collection_id
-
-    tag = create_tag(session=db_session, collection_id=collection_id)
-    image_1 = create_image(
-        session=db_session, collection_id=collection_id, file_path_abs="sample1.png"
-    )
-    image_2 = create_image(
-        session=db_session, collection_id=collection_id, file_path_abs="sample2.png"
-    )
-
-    tag_resolver.add_tag_to_sample(session=db_session, tag_id=tag.tag_id, sample=image_1.sample)
-    tag_resolver.add_tag_to_sample(session=db_session, tag_id=tag.tag_id, sample=image_2.sample)
-
-    tag_resolver.delete(session=db_session, tag_id=tag.tag_id)
-
-    tag_deleted = tag_resolver.get_by_id(session=db_session, tag_id=tag.tag_id)
-    assert tag_deleted is None
-    assert image_1.sample.tags == []
-    assert image_2.sample.tags == []
-
-
-def test_get_names_by_ids(db_session: Session) -> None:
-    collection = create_collection(session=db_session)
-    cid = collection.collection_id
-
-    tag_a = create_tag(session=db_session, collection_id=cid, tag_name="alpha")
-    tag_b = create_tag(session=db_session, collection_id=cid, tag_name="beta")
-
-    result = tag_resolver.get_names_by_ids(session=db_session, tag_ids=[tag_a.tag_id, tag_b.tag_id])
-    assert result == {tag_a.tag_id: "alpha", tag_b.tag_id: "beta"}
-
-
-def test_get_names_by_ids__empty(db_session: Session) -> None:
-    result = tag_resolver.get_names_by_ids(session=db_session, tag_ids=[])
-    assert result == {}
-
-
-def test_get_names_by_ids__unknown_id(db_session: Session) -> None:
-    result = tag_resolver.get_names_by_ids(session=db_session, tag_ids=[uuid4()])
-    assert result == {}
-
-
-def test_get_tags_by_sample(db_session: Session) -> None:
-    collection = create_collection(session=db_session)
-    cid = collection.collection_id
-
-    img_a = create_image(session=db_session, collection_id=cid, file_path_abs="a.png")
-    img_b = create_image(session=db_session, collection_id=cid, file_path_abs="b.png")
-
-    tag_1 = create_tag(session=db_session, collection_id=cid, tag_name="tag_1")
-    tag_2 = create_tag(session=db_session, collection_id=cid, tag_name="tag_2")
-
-    tag_resolver.add_tag_to_sample(session=db_session, tag_id=tag_1.tag_id, sample=img_a.sample)
-    tag_resolver.add_tag_to_sample(session=db_session, tag_id=tag_2.tag_id, sample=img_a.sample)
-    tag_resolver.add_tag_to_sample(session=db_session, tag_id=tag_1.tag_id, sample=img_b.sample)
-
-    result = tag_resolver.get_tags_by_sample(
-        session=db_session, tag_ids=[tag_1.tag_id, tag_2.tag_id]
-    )
-    assert result[img_a.sample_id] == {tag_1.tag_id, tag_2.tag_id}
-    assert result[img_b.sample_id] == {tag_1.tag_id}
-
-
-def test_get_tags_by_sample__empty(db_session: Session) -> None:
-    result = tag_resolver.get_tags_by_sample(session=db_session, tag_ids=[])
-    assert result == {}
-
-
-def test_get_tags_by_sample__no_memberships(db_session: Session) -> None:
-    collection = create_collection(session=db_session)
-    cid = collection.collection_id
-
-    tag = create_tag(session=db_session, collection_id=cid, tag_name="lonely")
-
-    result = tag_resolver.get_tags_by_sample(session=db_session, tag_ids=[tag.tag_id])
-    assert result == {}
-
-
-def test_get_or_create_sample_tag_by_name(db_session: Session) -> None:
-    collection = create_collection(session=db_session)
-    collection_id = collection.collection_id
-
-    # Create an existing tag
-    existing_tag = create_tag(
-        session=db_session, collection_id=collection_id, tag_name="existing_tag"
-    )
-
-    # Case 1: Get existing tag
-    result_tag = tag_resolver.get_or_create_sample_tag_by_name(
-        session=db_session, collection_id=collection_id, tag_name="existing_tag"
-    )
-    assert result_tag.tag_id == existing_tag.tag_id
-    assert result_tag.name == "existing_tag"
-
-    # Case 2: Create new tag
-    new_tag = tag_resolver.get_or_create_sample_tag_by_name(
-        session=db_session, collection_id=collection_id, tag_name="new_tag"
-    )
-    assert new_tag.tag_id != existing_tag.tag_id
-    assert new_tag.name == "new_tag"
-    assert new_tag.collection_id == collection_id
-    assert new_tag.kind == "sample"
-
-
-def test_add_tag_to_sample(db_session: Session) -> None:
-    collection = create_collection(session=db_session)
-    collection_id = collection.collection_id
-    tag = create_tag(session=db_session, collection_id=collection_id, kind="sample")
-    image = create_image(session=db_session, collection_id=collection_id)
-
-    # add sample to tag
-    tag_resolver.add_tag_to_sample(session=db_session, tag_id=tag.tag_id, sample=image.sample)
-
-    assert image.sample.tags.index(tag) == 0
-
-
-def test_remove_sample_from_tag(db_session: Session) -> None:
-    collection = create_collection(session=db_session)
-    collection_id = collection.collection_id
-    tag = create_tag(session=db_session, collection_id=collection_id, kind="sample")
-    image = create_image(session=db_session, collection_id=collection_id)
-
-    # add sample to tag
-    tag_resolver.add_tag_to_sample(session=db_session, tag_id=tag.tag_id, sample=image.sample)
-    assert len(image.sample.tags) == 1
-    assert image.sample.tags.index(tag) == 0
-
-    # remove sample to tag
-    tag_resolver.remove_tag_from_sample(session=db_session, tag_id=tag.tag_id, sample=image.sample)
-    assert len(image.sample.tags) == 0
-    with pytest.raises(ValueError, match="not in list"):
-        image.sample.tags.index(tag)
-
-
-def test_add_and_remove_sample_ids_to_tag_id(
-    db_session: Session,
-) -> None:
-    collection = create_collection(session=db_session)
-    collection_id = collection.collection_id
-    tag_1 = create_tag(
-        session=db_session,
-        collection_id=collection_id,
-        tag_name="tag_all",
-        kind="sample",
-    )
-    tag_2 = create_tag(
-        session=db_session,
-        collection_id=collection_id,
-        tag_name="tag_odd",
-        kind="sample",
-    )
-
-    total_samples = 10
-    images = []
-    for i in range(total_samples):
-        image = create_image(
-            session=db_session,
-            collection_id=collection_id,
-            file_path_abs=f"sample{i}.png",
-        )
-        images.append(image)
-
-    # add samples to tag_1
-    tag_resolver.add_sample_ids_to_tag_id(
-        session=db_session,
-        tag_id=tag_1.tag_id,
-        sample_ids=[sample.sample_id for sample in images],
-    )
-
-    # add every odd samples to tag_2
-    tag_resolver.add_sample_ids_to_tag_id(
-        session=db_session,
-        tag_id=tag_2.tag_id,
-        sample_ids=[sample.sample_id for i, sample in enumerate(images) if i % 2 == 1],
-    )
-
-    # ensure all samples were added to the correct tags
-    for i, image in enumerate(images):
-        assert tag_1 in image.sample.tags
-        if i % 2 == 1:
-            assert tag_2 in image.sample.tags
-
-    # ensure the correct number of samples were added to each tag
-    assert len(tag_1.samples) == total_samples
-    assert len(tag_2.samples) == total_samples / 2
-
-    # Remove the *same* even indexed samples we added earlier,
-    # but computed from the original `samples` list so ordering is stable.
-    tag_resolver.remove_sample_ids_from_tag_id(
-        session=db_session,
-        tag_id=tag_1.tag_id,
-        sample_ids=[s.sample_id for i, s in enumerate(images) if i % 2 == 0],
-    )
-
-    assert len(tag_1.samples) == total_samples / 2
-    assert len(tag_2.samples) == total_samples / 2
-
-    tag_1_samples_sorted = sorted(tag_1.samples, key=lambda s: s.sample_id)
-    tag_2_samples_sorted = sorted(tag_2.samples, key=lambda s: s.sample_id)
-    assert tag_1_samples_sorted == tag_2_samples_sorted
-
-
-def test_add_samples_to_tag_from_query__tags_only_query_matches(db_session: Session) -> None:
-    collection_id = create_collection(session=db_session).collection_id
-    tag = create_tag(session=db_session, collection_id=collection_id, kind="sample")
-    images = [
-        create_image(session=db_session, collection_id=collection_id, file_path_abs=f"s{i}.png")
-        for i in range(4)
-    ]
-    wanted = {images[0].sample_id, images[2].sample_id}
-
-    query = select(SampleTable.sample_id).where(col(SampleTable.sample_id).in_(wanted))
-    result = tag_resolver.add_samples_to_tag_from_query(
-        session=db_session, tag_id=tag.tag_id, sample_ids_query=query
-    )
-
-    # Exactly the query's matches are linked (correct sample/tag wiring, no over-tagging).
-    assert result is not None
-    assert {sample.sample_id for sample in tag.samples} == wanted
-
-
-def test_add_samples_to_tag_from_query__unknown_tag_returns_none(db_session: Session) -> None:
-    collection_id = create_collection(session=db_session).collection_id
-
-    query = select(SampleTable.sample_id).where(col(SampleTable.collection_id) == collection_id)
-    result = tag_resolver.add_samples_to_tag_from_query(
-        session=db_session, tag_id=uuid4(), sample_ids_query=query
-    )
-
-    assert result is None
-
-
-def test_add_and_remove_sample_ids_to_tag_id__twice_same_sample_ids(
-    db_session: Session,
-) -> None:
-    collection = create_collection(session=db_session)
-    collection_id = collection.collection_id
-    tag_1 = create_tag(
-        session=db_session,
-        collection_id=collection_id,
-        tag_name="tag_all",
-        kind="sample",
-    )
-
-    total_samples = 10
-    images = []
-    for i in range(total_samples):
-        image = create_image(
-            session=db_session,
-            collection_id=collection_id,
-            file_path_abs=f"sample{i}.png",
-        )
-        images.append(image)
-
-    # add samples to tag_1
-    tag_resolver.add_sample_ids_to_tag_id(
-        session=db_session,
-        tag_id=tag_1.tag_id,
-        sample_ids=[sample.sample_id for sample in images],
-    )
-
-    # adding the same samples to tag_1 does not create an error
-    tag_resolver.add_sample_ids_to_tag_id(
-        session=db_session,
-        tag_id=tag_1.tag_id,
-        sample_ids=[sample.sample_id for sample in images],
-    )
-
-    # ensure all samples were added once
-    assert len(tag_1.samples) == total_samples
-
-    # remove samples from
-    tag_resolver.remove_sample_ids_from_tag_id(
-        session=db_session,
-        tag_id=tag_1.tag_id,
-        sample_ids=[sample.sample_id for sample in images],
-    )
-    # removing the same samples to tag_1 does not create an error
-    tag_resolver.remove_sample_ids_from_tag_id(
-        session=db_session,
-        tag_id=tag_1.tag_id,
-        sample_ids=[sample.sample_id for sample in images],
-    )
-
-    # ensure all samples were removed again
-    assert len(tag_1.samples) == 0
